Fix pashov pre-audit suggestions

+++ src/interfaces/ISmoothlyPool.sol
@@ -1,9 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
 // Copyright 2022-2023 Smoothly Protocol LLC
-// SPDX License identifier: Apache-2.0
 pragma solidity 0.8.19;
 
 interface ISmoothlyPool {
-
   /// @notice Updates epoch number and Merkle root hashes 
   /// @param _withdrawalsRoot Merkle root hash for withdrawals
   /// @param _exitsRoot Merkle root hash for exits 
@@ -13,11 +12,11 @@ interface ISmoothlyPool {
     bytes32 _withdrawalsRoot, 
     bytes32 _exitsRoot,
     bytes32 _stateRoot,
-    uint _fee
+    uint256 _fee
   ) external;
 
   /// @dev Transfers ownership of the contract to a new account (`newOwner`).
-  /// Can only be called by the current owner.
+  /// Used to update PoolGovernance contract.
+  /// @param newOwner owner to transfer ownership to.
   function transferOwnership(address newOwner) external;
-
 }
